requirements.txt
Flask>=3.0.0
Pillow>=10.2.0
ImageHash>=4.3.1
opencv-python>=4.8.0
numpy>=1.24.0
gunicorn>=21.2.0
pywebview>=4.0
pyinstaller>=6.0
